requirements.txt
streamlit
langchain
faiss-cpu
sentence_transformers
transformers
peft
accelerate
huggingface_hub
pandas 
numpy 
python-dotenv
fastapi
langchain-core
llama-cpp-python
langchain_community
torch
pypdf
sentence-transformers